+++ accdfl/core/datasets/CIFAR10.py
@@ -2,13 +2,11 @@ import torch
 import torch.nn.functional as F
 import torchvision
 import torchvision.transforms as transforms
-from torch import nn
 from torch.utils.data import DataLoader
 
 from accdfl.core.datasets.Dataset import Dataset
 from accdfl.core.datasets.Partitioner import DataPartitioner, KShardDataPartitioner
 from accdfl.core.mappings.Mapping import Mapping
-from accdfl.core.models.Model import Model
 
 NUM_CLASSES = 10
 
@@ -212,93 +210,3 @@ class CIFAR10(Dataset):
         accuracy = float(correct) / float(example_number) * 100.0
         loss = total_loss / float(example_number)
         return accuracy, loss
-
-
-class CNN(Model):
-    """
-    Class for a CNN Model for CIFAR10
-
-    """
-
-    def __init__(self):
-        """
-        Constructor. Instantiates the CNN Model
-            with 10 output classes
-
-        """
-        super().__init__()
-        # 1.6 million params
-        self.conv1 = nn.Conv2d(3, 6, 5)
-        self.pool = nn.MaxPool2d(2, 2)
-        self.conv2 = nn.Conv2d(6, 16, 5)
-        self.fc1 = nn.Linear(16 * 5 * 5, 120)
-        self.fc2 = nn.Linear(120, 84)
-        self.fc3 = nn.Linear(84, NUM_CLASSES)
-
-    def forward(self, x):
-        """
-        Forward pass of the model
-
-        Parameters
-        ----------
-        x : torch.tensor
-            The input torch tensor
-
-        Returns
-        -------
-        torch.tensor
-            The output torch tensor
-
-        """
-        x = self.pool(F.relu(self.conv1(x)))
-        x = self.pool(F.relu(self.conv2(x)))
-        x = torch.flatten(x, 1)
-        x = F.relu(self.fc1(x))
-        x = F.relu(self.fc2(x))
-        x = self.fc3(x)
-        return x
-
-
-class LeNet(Model):
-    """
-    Class for a LeNet Model for CIFAR10
-    Inspired by original LeNet network for MNIST: https://ieeexplore.ieee.org/abstract/document/726791
-    """
-
-    def __init__(self):
-        """
-        Constructor. Instantiates the CNN Model
-            with 10 output classes
-
-        """
-        super().__init__()
-        self.conv1 = nn.Conv2d(3, 32, 5, padding="same")
-        self.pool = nn.MaxPool2d(2, 2)
-        self.gn1 = nn.GroupNorm(2, 32)
-        self.conv2 = nn.Conv2d(32, 32, 5, padding="same")
-        self.gn2 = nn.GroupNorm(2, 32)
-        self.conv3 = nn.Conv2d(32, 64, 5, padding="same")
-        self.gn3 = nn.GroupNorm(2, 64)
-        self.fc1 = nn.Linear(64 * 4 * 4, NUM_CLASSES)
-
-    def forward(self, x):
-        """
-        Forward pass of the model
-
-        Parameters
-        ----------
-        x : torch.tensor
-            The input torch tensor
-
-        Returns
-        -------
-        torch.tensor
-            The output torch tensor
-
-        """
-        x = self.pool(F.relu(self.gn1(self.conv1(x))))
-        x = self.pool(F.relu(self.gn2(self.conv2(x))))
-        x = self.pool(F.relu(self.gn3(self.conv3(x))))
-        x = torch.flatten(x, 1)
-        x = self.fc1(x)
-        return x
